requirements.txt
python-dotenv
requests
openai
openpyxl
Pillow
yfinance
matplotlib
